pyproject.toml
[tool.poetry]
name = "hananebot"
version = "0.1.0"
description = ""
authors = ["ShinonomeSetsuna <114331286+ShinonomeSetsuna@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
qq-botpy = "^1.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
